requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pymongo==4.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
rich==13.8.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
boto3
web3